+++ client/src/components/home/Form.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import styled from "styled-components";
-import PostFailed from "./PostFailed";
+import PostFailed from "../errors/PostFailed";
 
 const Form = (props) => {
   const { setNewTweetPost } = props;
